src/requirements.txt
amqp==5.0.6
anyjson==0.3.3
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pytz==2021.1
redis==3.5.3
requests==2.25.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
